+++ apphelpers/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Scroll Tech"""
 __email__ = "engg@stck.me"
-__version__ = "0.100.2"
+__version__ = "0.101.0"